better separation of the python and cpp version of the biot savart code
