+++ profiling/profiling.cpp
@@ -1,7 +1,10 @@
-#define FORCE_IMPORT_ARRAY
+//#define FORCE_IMPORT_ARRAY
 #include "xtensor/xnpy.hpp"
 #include "xtensor/xrandom.hpp"
-#include "biot_savart.h"
+
+#include "biot_savart_c.h"
+#include "biot_savart_vjp_c.h"
+
 #include <chrono>
 #include <iostream>
 #include <cstdlib>
@@ -55,6 +58,48 @@ void profile_biot_savart(int nsources, int ntargets, int nderivatives){
         << std::setw (19)<< clockcycles/interactions << std::endl;
 }
 
+void profile_biot_savart_vjp(int nsources, int ntargets, int nderivatives){ 
+    xt::xarray<double> points         = xt::random::randn<double>({ntargets, 3});
+    xt::xarray<double> gamma          = xt::random::randn<double>({nsources, 3});
+    xt::xarray<double> dgamma_by_dphi = xt::random::randn<double>({nsources, 3});
+    xt::xarray<double> v = xt::random::randn<double>({ntargets, 3});
+    xt::xarray<double> vgrad = xt::random::randn<double>({ntargets, 3, 3});
+
+    auto pointsx = vector_type(points.shape(0), 0);
+    auto pointsy = vector_type(points.shape(0), 0);
+    auto pointsz = vector_type(points.shape(0), 0);
+    for (int i = 0; i < points.shape(0); ++i) {
+        pointsx[i] = points(i, 0);
+        pointsy[i] = points(i, 1);
+        pointsz[i] = points(i, 2);
+    }
+
+    auto res_gamma = xt::xarray<double>::from_shape({points.shape(0), 3});
+    auto res_dgamma_by_dphi = xt::xarray<double>::from_shape({points.shape(0), 3});
+    auto res_grad_gamma = xt::xarray<double>::from_shape({points.shape(0), 3});
+    auto res_grad_dgamma_by_dphi = xt::xarray<double>::from_shape({points.shape(0), 3});
+
+
+    int n = int(1e8/(nsources*ntargets));
+
+    uint64_t tick = rdtsc();  // tick before
+    auto t1 = std::chrono::high_resolution_clock::now();
+    for (int i = 0; i < n; ++i) {
+        if(nderivatives == 0)
+            biot_savart_vjp_kernel<xt::xarray<double>, 0>(pointsx, pointsy, pointsz, gamma, dgamma_by_dphi, v, res_gamma, res_dgamma_by_dphi, vgrad, res_grad_gamma, res_grad_dgamma_by_dphi);
+        else
+            biot_savart_vjp_kernel<xt::xarray<double>, 1>(pointsx, pointsy, pointsz, gamma, dgamma_by_dphi, v, res_gamma, res_dgamma_by_dphi, vgrad, res_grad_gamma, res_grad_dgamma_by_dphi);
+    }
+    auto t2 = std::chrono::high_resolution_clock::now();
+    auto clockcycles = rdtsc() - tick;
+    double simdtime = std::chrono::duration_cast<std::chrono::milliseconds>( t2 - t1 ).count();
+    double interactions = points.shape(0) * gamma.shape(0) * n;
+    std::cout << std::setw (10) << nsources*ntargets 
+        << std::setw (13) << simdtime/n 
+        << std::setw (19) << std::setprecision(5) << (interactions/(1e9 * simdtime/1000.)) 
+        << std::setw (19)<< clockcycles/interactions << std::endl;
+}
+
 
 int main() {
     for(int nd=0; nd<3; nd++) {
@@ -62,6 +107,13 @@ int main() {
         std::cout << "         N" << " Time (in ms)" << " Gigainteractions/s" << " cycles/interaction" << std::endl;
         for(int nst=10; nst<=10000; nst*=10)
             profile_biot_savart(nst, nst, nd);
+    }
+
+    for(int nd=0; nd<2; nd++) {
+        std::cout << "Number of derivatives: " << nd << std::endl;
+        std::cout << "         N" << " Time (in ms)" << " Gigainteractions/s" << " cycles/interaction" << std::endl;
+        for(int nst=10; nst<=10000; nst*=10)
+            profile_biot_savart_vjp(nst, nst, nd);
     }
     return 0;
 }
