+++ src/simsgeopp/biot_savart_vjp_c.h
@@ -1,0 +1,5 @@
+#pragma once
+#include "simdhelpers.h"
+
+template<class T, int derivs>
+void biot_savart_vjp_kernel(vector_type& pointsx, vector_type& pointsy, vector_type& pointsz, T& gamma, T& dgamma_by_dphi, T& v, T& res_gamma, T& res_dgamma_by_dphi, T& vgrad, T& res_grad_gamma, T& res_grad_dgamma_by_dphi);
